6. Разнес utils по файлам
