+++ src/utils/common.js
@@ -1,0 +1,3 @@
+const capitalizeFirstLetter = (inputString) => inputString[0].toUpperCase() + inputString.slice(1);
+
+export { capitalizeFirstLetter };
