+++ src/utils/event.js
@@ -1,13 +1,7 @@
 import dayjs from 'dayjs';
 import duration from 'dayjs/plugin/duration';
-import IsSameOrAfter from 'dayjs/plugin/isSameOrAfter';
-import isSameOrBefore from 'dayjs/plugin/isSameOrBefore';
 dayjs.extend(duration);
-dayjs.extend(IsSameOrAfter);
-dayjs.extend(isSameOrBefore);
 
-
-const capitalizeFirstLetter = (inputString) => inputString[0].toUpperCase() + inputString.slice(1);
 
 const reformatDate = (data) => (
   {
@@ -35,13 +29,4 @@ const calculateDuration = (startDate, endDate) => {
 };
 
 
-const isFutureEvent = (startDate) => dayjs(startDate).isAfter(dayjs(), 'date');
-
-const isPastEvent = (endDate) => dayjs(endDate).isBefore(dayjs(), 'date');
-
-const isPresentEvent = (startDate, endDate) =>
-  dayjs(startDate).isSameOrBefore(dayjs(), 'date') &&
-  dayjs(endDate).isSameOrAfter(dayjs(), 'date');
-
-
-export { capitalizeFirstLetter, reformatDate, isFutureEvent, isPastEvent, isPresentEvent, calculateDuration };
+export { reformatDate, calculateDuration };
